ex4.19 working again, 4.21 done

+++ Part4/Bloglist/controllers/blogs.js
@@ -2,7 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-const { tokenExtractor } = require('../utils/middleware')
+//const { tokenExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -22,20 +22,19 @@ blogsRouter.get('/:id', (request, response, next) => {
     })
     .catch(error => next(error))
 })
-/*const getTokenFrom = request => {
+
+const getTokenFrom = request => {
   const authorization = request.get('authorization')
   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
     return authorization.substring(7)
   }
   return null
 }
-*/
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const authorization = request.get('authorization')
-  const token = tokenExtractor(authorization)
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
   if (!token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
@@ -45,19 +44,32 @@ blogsRouter.post('/', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    user: user,
-    likes: body.likes,
+    user: user._id,
+    likes: body.likes === undefined ? 0 : body.likes
   })
 
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
+  response.json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
 
-  response.status(204).end()
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    return response.status(401).json({ error: 'wrong token' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -70,8 +82,8 @@ blogsRouter.put('/:id', async (request, response) => {
     likes: body.likes
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-  response.json(updatedBlog.toJSON())
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true, runValidators: true, upsert: true, context: 'query' })
+  response.status(200).json(updatedBlog.toJSON()) //.json(result)
 })
 
 module.exports = blogsRouter
